Introduce pformat method and refactor to use in dbg
