+++ debug/__init__.py
@@ -1,3 +1,4 @@
 from ._debug import CONFIG, dbg
+from ._format import pformat
 
-__all__ = ["dbg", "CONFIG"]
+__all__ = ["dbg", "pformat", "CONFIG"]
