+++ debug/_format.py
@@ -6,10 +6,9 @@ import ast
 from collections import ChainMap, Counter, UserDict, UserList, defaultdict, deque
 from collections.abc import Callable, Collection, ItemsView, Iterable, KeysView, ValuesView
 import dataclasses
-from dataclasses import dataclass, field
-import os
+from dataclasses import dataclass
+import functools
 import re
-import sys
 import textwrap
 from types import MethodWrapperType
 from typing import (
@@ -25,8 +24,7 @@ import unicodedata
 
 from wcwidth import wcswidth
 
-from ._code import highlight_code
-from ._config import DbgConfig, pytest_enabled
+from ._code import highlight_code, validate_style
 
 frozendict: type[Any]
 try:
@@ -72,16 +70,16 @@ class BaseFormat(abc.ABC):
     def length(self) -> int | None: ...
 
     @abc.abstractmethod
-    def _format(self, used_width: int, highlight: bool, config: FormatterConfig) -> str: ...
+    def _format(self, highlight_now: bool, config: FormatterConfig) -> str: ...
 
-    def _highlight_code(self, highlight: bool, text: str) -> str:
-        if highlight and self._highlight:
-            return highlight_code(text)
+    def _highlight_code(self, style: str | None, text: str) -> str:
+        if style is not None and self._highlight_subitems:
+            return highlight_code(text, style)
         return text
 
     @property
     @abc.abstractmethod
-    def _highlight(self) -> bool: ...
+    def _highlight_subitems(self) -> bool: ...
 
     @classmethod
     def total_length(cls, objs: Iterable[BaseFormat]) -> int | None:
@@ -260,16 +258,20 @@ class SequenceFormat(BaseFormat, abc.ABC):
         open, close = self.parentheses
         return len(open) + len(close)
 
-    def _format(self, used_width: int, highlight: bool, config: FormatterConfig) -> str:
+    def _format(self, highlight_now: bool, config: FormatterConfig) -> str:
         if self._objs is None:
             return self._empty_format(config)
         length = self.length()
         if len(self._objs) == 0 or (
             length is not None
-            and (config.terminal_width is None or length <= config.terminal_width - used_width)
+            and (config.remaining_width is None or length <= config.remaining_width)
         ):
-            return self._highlight_code(highlight, self._flat_format(self._objs, config))
-        return self._highlight_code(highlight, self._nested_format(self._objs, used_width, config))
+            return self._highlight_code(
+                config.get_style(highlight_now), self._flat_format(self._objs, config)
+            )
+        return self._highlight_code(
+            config.get_style(highlight_now), self._nested_format(self._objs, config)
+        )
 
     @property
     def parentheses(self) -> tuple[str, str]:
@@ -294,23 +296,22 @@ class SequenceFormat(BaseFormat, abc.ABC):
         config = config.flatten()
         return (
             open
-            + ", ".join(
-                obj._format(len(open) + len(close), not self._highlight, config) for obj in objs
-            )
+            + ", ".join(obj._format(not self._highlight_subitems, config) for obj in objs)
             + ("," if self._extra_trailing_comma else "")
             + close
         )
 
-    def _nested_format(
-        self, objs: list[BaseFormat], used_width: int, config: FormatterConfig
-    ) -> str:
+    def _nested_format(self, objs: list[BaseFormat], config: FormatterConfig) -> str:
         """Return a formatted sequence across multiple lines."""
         open, close = self.parentheses
         config = config.indent()
         return (
             f"{open}\n"
             + textwrap.indent(
-                "\n".join(f"{obj._format(1, not self._highlight, config)}," for obj in objs),
+                "\n".join(
+                    f"{obj._format(not self._highlight_subitems, config.use_only(1))},"
+                    for obj in objs
+                ),
                 prefix=config.get_indent(),
             )
             + f"\n{close}"
@@ -320,11 +321,11 @@ class SequenceFormat(BaseFormat, abc.ABC):
     @abc.abstractmethod
     def _parentheses(self) -> tuple[str, str]: ...
 
-    @property
-    def _highlight(self) -> bool:
+    @functools.cached_property
+    def _highlight_subitems(self) -> bool:
         if self._objs is None:
             return True
-        return all(obj._highlight for obj in self._objs)
+        return all(obj._highlight_subitems for obj in self._objs)
 
 
 class SquareSequenceFormat(SequenceFormat):
@@ -366,25 +367,25 @@ class PairFormat(BaseFormat):
         self._value = value
         self._length = self.add(self.add(self._key.length(), self._value.length()), 2)
 
-    def _format(self, used_width: int, highlight: bool, config: FormatterConfig) -> str:
-        key_format = self._key._format(1, not self._highlight, config) + ":"
+    def _format(self, highlight: bool, config: FormatterConfig) -> str:
+        key_format = self._key._format(not self._highlight_subitems, config.use_only(1)) + ":"
         lines = key_format.splitlines()
         offset_width = self.len(lines[-1]) + 1
         value_format = " " + self._value._format(
-            offset_width + used_width, not self._highlight, config
+            not self._highlight_subitems, config.use_extra(offset_width)
         )
         if isinstance(self._value, ItemFormat) and self._value.length() is None:
             value_format = textwrap.indent(
                 value_format, prefix=" " * offset_width, predicate=not_first()
             )
-        return self._highlight_code(highlight, key_format + value_format)
+        return self._highlight_code(config.get_style(highlight), key_format + value_format)
 
     def length(self) -> int | None:
         return self._length
 
-    @property
-    def _highlight(self) -> bool:
-        return self._key._highlight and self._value._highlight
+    @functools.cached_property
+    def _highlight_subitems(self) -> bool:
+        return self._key._highlight_subitems and self._value._highlight_subitems
 
 
 class AttrFormat(BaseFormat):
@@ -393,23 +394,23 @@ class AttrFormat(BaseFormat):
         self._value = value
         self._length = self.add(1 + len(self._attr), self._value.length())
 
-    def _format(self, used_width: int, highlight: bool, config: FormatterConfig) -> str:
+    def _format(self, highlight: bool, config: FormatterConfig) -> str:
         attr_format = self._attr + "="
         value_format = self._value._format(
-            used_width + len(attr_format), not self._highlight, config
+            not self._highlight_subitems, config.use_extra(len(attr_format))
         )
         if isinstance(self._value, ItemFormat) and self._value.length() is None:
             value_format = textwrap.indent(
                 value_format, prefix=" " * len(attr_format), predicate=not_first()
             )
-        return self._highlight_code(highlight, attr_format + value_format)
+        return self._highlight_code(config.get_style(highlight), attr_format + value_format)
 
     def length(self) -> int | None:
         return self._length
 
-    @property
-    def _highlight(self) -> bool:
-        return self._value._highlight
+    @functools.cached_property
+    def _highlight_subitems(self) -> bool:
+        return self._value._highlight_subitems
 
 
 class ItemFormat(BaseFormat):
@@ -421,15 +422,15 @@ class ItemFormat(BaseFormat):
             return None
         return self.len(self.repr)
 
-    def _format(self, used_width: int, highlight: bool, config: FormatterConfig) -> str:
-        return self._highlight_code(highlight, self.repr)
+    def _format(self, highlight_now: bool, config: FormatterConfig) -> str:
+        return self._highlight_code(config.get_style(highlight_now), self.repr)
 
     @property
     def multiline(self) -> bool:
         return "\n" in self.repr
 
-    @property
-    def _highlight(self) -> bool:
+    @functools.cached_property
+    def _highlight_subitems(self) -> bool:
         return strip_ansi(self.repr) == self.repr
 
 
@@ -659,66 +660,86 @@ BaseFormat.KNOWN_EXTRA_CLASSES = (
 
 @dataclass(repr=False, kw_only=True, frozen=True)
 class FormatterConfig:
-    DEFAULT_WIDTH: ClassVar[int] = 80
+    indent_width: int
+    width_pair: None | tuple[int, int]
+    style: str | None
 
-    @staticmethod
-    def _get_terminal_width() -> int:
-        width = FormatterConfig.DEFAULT_WIDTH
-        try:
-            width, _ = os.get_terminal_size(sys.stderr.fileno())
-        except OSError:
-            if pytest_enabled():
-                stderr = sys.__stderr__
-                if stderr is not None:
-                    try:
-                        width, _ = os.get_terminal_size(stderr.fileno())
-                    except OSError:
-                        ...
-        if width < FormatterConfig.DEFAULT_WIDTH / 2:
-            width = FormatterConfig.DEFAULT_WIDTH
-        return width
+    def __post_init__(self) -> None:
+        if self.style is not None:
+            validate_style(self.style)
 
-    indent_width: int = 2
-    terminal_width: None | int = field(default_factory=_get_terminal_width)
-    color: bool
+    @property
+    def remaining_width(self) -> int | None:
+        if self.width_pair is None:
+            return None
+        used_width, terminal_width = self.width_pair
+        return terminal_width - used_width
 
     def indent(self) -> Self:
-        return type(self)(
-            indent_width=self.indent_width,
-            terminal_width=BaseFormat.add(
-                self.terminal_width,
-                -self.indent_width,
-            ),
-            color=self.color,
+        if self.width_pair is None:
+            return self
+        used_width, terminal_width = self.width_pair
+        return self.replace(
+            width_pair=(used_width, terminal_width - self.indent_width),
         )
 
+    def replace(self, **kwargs: Any) -> Self:
+        return dataclasses.replace(self, **kwargs)
+
     def flatten(self) -> Self:
-        return type(self)(indent_width=self.indent_width, terminal_width=None, color=self.color)
+        return self.replace(width_pair=None)
 
     def get_indent(self) -> str:
         return " " * self.indent_width
 
-    @classmethod
-    def _from_config(cls, config: DbgConfig) -> Self:
-        return cls(indent_width=config.indent, color=config.color)
-
-
-class Formatter:
-    def __init__(self, config: FormatterConfig) -> None:
-        self._config = config
-
-    def format(self, obj: Any, *, initial_width: int) -> str:
-        formatted_obj = BaseFormat._from(obj, set())
-        text = formatted_obj._format(
-            initial_width, highlight=self._config.color, config=self._config
+    def _update_used_space(self, update: Callable[[int], int]) -> Self:
+        if self.width_pair is None:
+            return self
+        used_width, terminal_width = self.width_pair
+        return self.replace(
+            width_pair=(update(used_width), terminal_width),
         )
-        terminal_width = self._config.terminal_width
-        if isinstance(formatted_obj, ItemFormat) and text:
-            max_len = max(map(BaseFormat.len, text.splitlines()))
-            if terminal_width is not None and max_len + initial_width > terminal_width:
-                text = "\n" + text
-                if initial_width + 1 <= terminal_width:
-                    text = "\u23ce" + text
-            else:
-                text = textwrap.indent(text, " " * initial_width, predicate=not_first())
-        return text
+
+    def use_extra(self, space: int) -> Self:
+        return self._update_used_space(lambda used_width: used_width + space)
+
+    def use_only(self, space: int) -> Self:
+        return self._update_used_space(lambda used_width: space)
+
+    @property
+    def color(self) -> bool:
+        return self.style is not None
+
+    def get_style(self, enabled: bool) -> str | None:
+        if not enabled:
+            return None
+        return self.style
+
+
+def pformat(
+    obj: Any,
+    /,
+    *,
+    width: int | None = 80,
+    style: str | None = None,
+    indent: int = 2,
+    prefix: str = "",
+) -> str:
+    *_, last_line = prefix.rsplit("\n", maxsplit=1)
+    initial_offset = BaseFormat.len(last_line)
+    if width is None:
+        width_pair = None
+    else:
+        width_pair = (initial_offset, width)
+    config = FormatterConfig(width_pair=width_pair, indent_width=indent, style=style)
+    formatted_obj = BaseFormat._from(obj, set())
+    text = formatted_obj._format(highlight_now=True, config=config)
+    if prefix and isinstance(formatted_obj, ItemFormat) and text:
+        max_len = max(map(BaseFormat.len, text.splitlines()))
+        if config.remaining_width is not None and max_len > config.remaining_width:
+            text = "\n" + text
+            if config.remaining_width >= 1:
+                text = "\u23ce" + text
+        else:
+            text = textwrap.indent(text, " " * initial_offset, predicate=not_first())
+    return prefix + text
